+++ src/lv_core/lv_core.mk
@@ -4,7 +4,6 @@ CSRCS += lv_disp.c
 CSRCS += lv_obj.c
 CSRCS += lv_refr.c
 CSRCS += lv_style.c
-CSRCS += lv_debug.c
 
 DEPPATH += --dep-path $(LVGL_DIR)/$(LVGL_DIR_NAME)/src/lv_core
 VPATH += :$(LVGL_DIR)/$(LVGL_DIR_NAME)/src/lv_core
