+++ src/lv_misc/lv_misc.mk
@@ -14,6 +14,7 @@ CSRCS += lv_utils.c
 CSRCS += lv_async.c
 CSRCS += lv_printf.c
 CSRCS += lv_bidi.c
+CSRCS += lv_debug.c
 
 DEPPATH += --dep-path $(LVGL_DIR)/$(LVGL_DIR_NAME)/src/lv_misc
 VPATH += :$(LVGL_DIR)/$(LVGL_DIR_NAME)/src/lv_misc
