+++ src/lv_widgets/lv_btn.c
@@ -12,7 +12,7 @@
 
 #include <string.h>
 #include "../lv_core/lv_group.h"
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_draw/lv_draw.h"
 #include "../lv_themes/lv_theme.h"
 #include "../lv_misc/lv_area.h"
