+++ src/lv_widgets/lv_checkbox.c
@@ -9,7 +9,7 @@
 #include "lv_checkbox.h"
 #if LV_USE_CHECKBOX != 0
 
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_core/lv_group.h"
 #include "../lv_themes/lv_theme.h"
