+++ src/lv_widgets/lv_cpicker.c
@@ -10,7 +10,7 @@
 #include "lv_cpicker.h"
 #if LV_USE_CPICKER != 0
 
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_draw/lv_draw_arc.h"
 #include "../lv_themes/lv_theme.h"
 #include "../lv_core/lv_indev.h"
