+++ src/lv_widgets/lv_label.c
@@ -9,7 +9,7 @@
 #include "lv_label.h"
 #if LV_USE_LABEL != 0
 #include "../lv_core/lv_obj.h"
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_core/lv_group.h"
 #include "../lv_draw/lv_draw.h"
 #include "../lv_misc/lv_color.h"
