+++ src/lv_widgets/lv_line.c
@@ -9,7 +9,7 @@
 #include "lv_line.h"
 
 #if LV_USE_LINE != 0
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_draw/lv_draw.h"
 #include "../lv_misc/lv_math.h"
 #include "../lv_themes/lv_theme.h"
