+++ src/lv_widgets/lv_linemeter.c
@@ -9,7 +9,7 @@
 #include "lv_linemeter.h"
 #if LV_USE_LINEMETER != 0
 
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_draw/lv_draw.h"
 #include "../lv_themes/lv_theme.h"
 #include "../lv_core/lv_group.h"
