+++ src/lv_widgets/lv_list.c
@@ -9,7 +9,7 @@
 #include "lv_list.h"
 #if LV_USE_LIST != 0
 
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_core/lv_group.h"
 #include "../lv_themes/lv_theme.h"
 #include "../lv_misc/lv_anim.h"
