+++ src/lv_widgets/lv_objmask.c
@@ -7,7 +7,7 @@
  *      INCLUDES
  *********************/
 #include "lv_objmask.h"
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_draw/lv_draw.h"
 #include "../lv_themes/lv_theme.h"
