+++ src/lv_widgets/lv_page.c
@@ -9,7 +9,7 @@
 #include "../lv_widgets/lv_page.h"
 #if LV_USE_PAGE != 0
 
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_core/lv_group.h"
 #include "../lv_draw/lv_draw.h"
 #include "../lv_themes/lv_theme.h"
