+++ src/lv_widgets/lv_table.c
@@ -9,7 +9,7 @@
 #include "lv_table.h"
 #if LV_USE_TABLE != 0
 
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_core/lv_indev.h"
 #include "../lv_misc/lv_txt.h"
 #include "../lv_misc/lv_math.h"
