+++ src/lv_widgets/lv_tabview.c
@@ -10,7 +10,7 @@
 #if LV_USE_TABVIEW != 0
 
 #include "lv_btnmatrix.h"
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_themes/lv_theme.h"
 #include "../lv_misc/lv_anim.h"
 #include "../lv_core/lv_disp.h"
