+++ src/lv_widgets/lv_tileview.c
@@ -12,7 +12,7 @@
 #include <stdbool.h>
 #include "lv_cont.h"
 #include "../lv_misc/lv_math.h"
-#include "../lv_core/lv_debug.h"
+#include "../lv_misc/lv_debug.h"
 #include "../lv_themes/lv_theme.h"
 
 /*********************
